начать реализовывать логику сервисов ItemService и UserService для работы с репозиториями

+++ src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -5,13 +5,13 @@ import ru.practicum.shareit.item.dto.ItemDto;
 import java.util.List;
 
 public interface ItemService {
-    ItemDto createItem(ItemDto itemDto, int userId);
+    ItemDto createItem(ItemDto itemDto, long userId);
 
-    ItemDto updateItem(ItemDto itemDto, int itemId, int userId);
+    ItemDto updateItem(ItemDto itemDto, long itemId, long userId);
 
-    ItemDto getItemDtoById(int itemId);
+    ItemDto getItemDtoById(long itemId);
 
-    List<ItemDto> getItemsByUserId(int userId);
+    List<ItemDto> getItemsByUserId(long userId);
 
     List<ItemDto> searchItem(String text);
 }
